Eliminate legacy code in organize.py
